Made the OpenID login a little friendlier
I tried out an OpenID jquery plugin, but it was really unpolished and
polluted the namespace.  On top of that, I couldn't get it to work.  I
decided that rather than messing wiht it, I'll just provide google &
yahoo links for now.  Most folks have one or the other.  Because they
use a fixed endpoint for all users, this was easier to do than any of
the others.  In the future, I hope to support all kinds of OpenID
providers with UI, but for now, if users of other services want to use
OpenID, they'll have to find their OpenID URL on their own.