Suggestion error messages & creation improvements
When suggestions are created, the currently logged in user is assigned
to the current suggestion.  If no user is logged in, they will get an
error.  That error needs some help.  At any rate, it prevents anonymous
suggestions, and displays errors when people fill out the form wrong.
Also, I tweaked some of the CSS to make it look a little shinier.