Solved 'do we know you?' problem.
We were redirecting everyone to the "tell us about yourself" page
instead of just folks we don't have a name for.  This was because we
were assigning a nickname of 'Guest_#{id}' to folks without a name.  By
switching to 'nickname' for this functionality, the raw 'name'
functionality was preserved.  Also, this helps differentiate between
'nickname' and 'fullname' in the rest of the codebase.